requirements.txt
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.2
django-bootstrap3==12.0.3
django-registration==3.0
django-tinymce==2.8.0
idna==2.9
Markdown==3.0
martor==1.4.7
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
